♻️ Refactor: The KakaoMap component #24
