+++ src/components/KakaoMap/KakaoMap.jsx
@@ -1,8 +1,9 @@
 import React, { useState, useEffect } from 'react';
-import { CustomOverlayMap, Map, MapMarker } from 'react-kakao-maps-sdk';
+import { Map, MapMarker } from 'react-kakao-maps-sdk';
 import { getUserLocation } from '../../utils/getUserLocation';
-import { Link } from 'react-router-dom';
 import { useRef } from 'react';
+import MyLocationMarker from './MyLocationMarker';
+import MapProductMarker from './MapProductMarker';
 
 const KakaoMap = ({ level, mode, productList, selectedProduct }) => {
   const [location, setLocation] = useState({ lat: null, lng: null }); // 유저의 중심 위치를 위한 상태
@@ -30,10 +31,6 @@ const KakaoMap = ({ level, mode, productList, selectedProduct }) => {
     }
   }, [selectedProduct, productList]);
 
-  const handleClickProduct = (product) => {
-    setProductInfo(product);
-  };
-
   const handleClickMap = () => {
     setProductInfo(null);
   };
@@ -46,70 +43,16 @@ const KakaoMap = ({ level, mode, productList, selectedProduct }) => {
         style={{ width: '100%', height: '100%' }}
         onClick={handleClickMap}
       >
-        <MapMarker
-          position={location} // ip 기반으로 현재 내 위치 지정
-          image={{
-            src: 'https://cdn-static.zep.us/static/assets/baked-avartar-images/10-35-3-253.png', // 프로필 사진 들어갈 예정
-            size: {
-              width: 39,
-              height: 39,
-            },
-            options: {
-              offset: {
-                x: 27,
-                y: 69,
-              },
-            },
-          }}
-        ></MapMarker>
-        {mode === 'productList' &&
-          productList.map((product) => (
-            <div key={product.id}>
-              <MapMarker
-                position={product.location}
-                onClick={() => handleClickProduct(product)}
-              />
-              {productInfo && productInfo.id === product.id && (
-                <CustomOverlayMap
-                  position={product.location}
-                  yAnchor={1.4}
-                  clickable={true}
-                >
-                  <Link
-                    to={`/product/detail/${product.id}`}
-                    style={{
-                      pointerEvents: 'auto',
-                      display: 'flex',
-                      flexDirection: 'column',
-                      alignItems: 'center',
-                      justifyContent: 'center',
-                      padding: '20px',
-                      backgroundColor: '#fff',
-                      border: '2px solid rgb(85 204 201)',
-                      borderRadius: '15px',
-                      color: '#000',
-                      textAlign: 'center',
-                      minWidth: '120px',
-                    }}
-                  >
-                    <div style={{ fontWeight: 'bold', marginBottom: '5px' }}>
-                      {product.name}
-                    </div>
-                    <div style={{ color: '#666', marginBottom: '5px' }}>
-                      {Number(product.price).toLocaleString()}원
-                    </div>
-                    <div style={{ fontSize: '12px', color: '#888' }}>
-                      {product.users.name}{' '}
-                      {new Date(product.createdAt)
-                        .toISOString()
-                        .slice(2, 10)
-                        .replace(/-/g, '.')}
-                    </div>
-                  </Link>
-                </CustomOverlayMap>
-              )}
-            </div>
-          ))}
+        {mode === 'productList' && (
+          <>
+            <MyLocationMarker location={location} />
+            <MapProductMarker
+              productList={productList}
+              productInfo={productInfo}
+              setProductInfo={setProductInfo}
+            />
+          </>
+        )}
       </Map>
     </div>
   );
